Eliminate synchronization conflicts between tsm and tsm2, and between iterator v1 and iterator v2.
